This is a squashed commit for test: Optimize visibility server as on 22 may at 12:41 IST
